+++ src/main/kotlin/com/github/okafke/poryscriptidea/lsp/PsLanguageServerFactory.kt
@@ -1,0 +1,23 @@
+package com.github.okafke.poryscriptidea.lsp
+
+import com.intellij.openapi.project.Project
+import com.redhat.devtools.lsp4ij.LanguageServerFactory
+import com.redhat.devtools.lsp4ij.client.LanguageClientImpl
+import com.redhat.devtools.lsp4ij.installation.ServerInstaller
+import com.redhat.devtools.lsp4ij.server.StreamConnectionProvider
+
+
+class PsLanguageServerFactory : LanguageServerFactory {
+    override fun createConnectionProvider(project: Project): StreamConnectionProvider {
+        return PsLanguageServer(project)
+    }
+
+    override fun createLanguageClient(project: Project): LanguageClientImpl {
+        return PsLanguageClient(project)
+    }
+
+    override fun createServerInstaller(): ServerInstaller {
+        return PsLanguageServerInstaller()
+    }
+
+}